Remove Ethers.RPC module
The Ethers.RPC module tried to serve a purpose as a facade between
Ethers functions and actual Json RPC implementation. The issue raises
when this module starts being used from the external dependencies and
suddenly they realize some of the logic for Ethers has leaked into this
module.

To address this and also to simplify, Ethers.RPC is removed and Ethers
module will make direct calls to the underlying RPC library
(Ethereumex.HttpClient by default).

Closes #34